Streamline fflogs.py
To separate the authentication and retrieval logic, it was necessary to introduce state to reduce requests to the API. This commit cleans up the previous commit that began outlining the class structure for the file. Superfluous commented code is deleted, the old implementation was removed, and the funtions tidied up.